fix: 32bit -> 16bit

+++ src/modules/display.cpp
@@ -39,7 +39,7 @@ void drawOilTemperatureTopBar(M5Canvas& canvas, float oilTemp, int maxOilTemp)
 
     if (oilTemp >= MIN_TEMP) {
         int barWidth = static_cast<int>(W * (oilTemp - MIN_TEMP) / RANGE);
-        uint32_t barColor = (oilTemp >= ALERT_TEMP) ? COLOR_RED : COLOR_WHITE;
+        uint16_t barColor = (oilTemp >= ALERT_TEMP) ? COLOR_RED : COLOR_WHITE;
         canvas.fillRect(X, Y, barWidth, H, barColor);
     }
